Add files via upload
on_ready makes a database+collection per guild/server that the bot is in. So, now it can be added to other places.  #5 still isn't completely fixed, I haven't added the cog listener but this commit basically preps for this.  Result: When the bot runs for the first time it makes a database/collection for each server it's in and for every prediction it calls the user's respective DB. Also exported some functions to Functions.py