Add files via upload
I changed the addPts() function to voiceChannelCheck(). Now, this function at line 84 is threading and works properly to check through every guild/server to see if there are people in the voice channel. If there are then it finds their collection and adds to their points from a random range of 90-125.